+++ backend/controllers/templatesController/templateCT2Controller.js
@@ -1,35 +1,46 @@
-import TemplateCT2 from "../models/TemplateModelCT2.js";
+import TemplateCT2 from "../../models/TemplateModelCT2.js";
 
 const getTemplates_CT2 = async (req, res) => {
   const templates = await TemplateCT2.find();
-  console.log("all - templates ct2",templates);
-  try{
+  console.log("all - templates ct2", templates);
+  try {
     res.json(templates);
-  }
-  catch(err){
-    consolee.log(`Error : ${err}`)
+  } catch (err) {
+    consolee.log(`Error : ${err}`);
   }
 };
+
 const createTemplate_CT2 = async (req, res) => {
-  // name, presion,pulse,temperature,FR,SAC,Obs
+  // client
+  // PC
+  // template id_
+  // creator
+  // Obs
+
   const template = new TemplateCT2(req.body);
-  template.creator= req.user._id;
-  // template.idTemplate = req.idTemplate;
+  template.creator = req.user._id;
+
   try {
     const CT2_Save = await template.save();
     res.json(CT2_Save);
-
   } catch (err) {
     console.log(`Error : ${err}`);
   }
 };
 
-const getSubSectionTemplate_CT2 = async (req, res) => {};
-
 const getTemplate_CT2 = async (req, res) => {
-  console.log("getClient : req.params",req.params);
+  const { id } = req.params;
+  const template = await TemplateCT2.findById(id);
+
+  console.log("TemplateModel : ", template);
+
+  if (!template) {
+    const err = new Error("No encontrado.");
+    return res.status(404).json({ msg: err.message });
+  }
+
   try {
-    res.json(req.body);
+    res.json(template);
   } catch (err) {
     console.log(`Error : ${err}`);
   }
@@ -39,16 +50,16 @@ const editTemplate_CT2 = async (req, res) => {
   const { id } = req.params;
 
   const template = await TemplateCT2.findById(id);
-  console.log("edit - templates C6", template);
+  // console.log("edit - templates C6", template);
 
-  if(!template){
-    const err = new Error('No encontrado.');
-    return res.status(404).json({msg:err.message})
+  if (!template) {
+    const err = new Error("No encontrado.");
+    return res.status(404).json({ msg: err.message });
   }
 
-  if(template.creator.toString() !== req.user._id.toString()){
-    const err = new Error('No Válido.');
-    return res.status(404).json({msg:err.message})
+  if (template.creator.toString() !== req.user._id.toString()) {
+    const err = new Error("No Válido.");
+    return res.status(404).json({ msg: err.message });
   }
   // other user : aprobación / regla de negocio
 
@@ -56,12 +67,11 @@ const editTemplate_CT2 = async (req, res) => {
   template.Obs = req.body.Obs || template.Obs;
 
   try {
-    const editTemplate =  await template.save();
+    const editTemplate = await template.save();
     res.json(editTemplate);
   } catch (err) {
     console.log(`Error : ${err}`);
   }
-
 };
 
 const deleteTemplate_CT2 = async (req, res) => {
@@ -69,9 +79,9 @@ const deleteTemplate_CT2 = async (req, res) => {
 
   const template = await TemplateCT2.findById(id);
 
-  if(!template){
-    const err = new Error('No encontrado.');
-    return res.status(404).json({msg:err.message})
+  if (!template) {
+    const err = new Error("No encontrado.");
+    return res.status(404).json({ msg: err.message });
   }
 
   // if(template.creator.toString() !== req.user._id.toString()){
@@ -80,29 +90,24 @@ const deleteTemplate_CT2 = async (req, res) => {
   // }
 
   // other user : aprobación / regla de negocio
-  if(req.user.rol.toString() !== 'SUPERADMIN'){
-    const err = new Error('No Válido.');
-    return res.status(404).json({msg:err.message});
+  if (req.user.rol.toString() !== "SUPERADMIN") {
+    const err = new Error("No Válido.");
+    return res.status(404).json({ msg: err.message });
   }
 
   try {
     await template.deleteOne();
-    res.json({msg:'Registro Eliminado.'});
+    res.json({ msg: "Registro Eliminado." });
   } catch (err) {
     console.log(`Error : ${err}`);
-  };
-}
+  }
+};
 
-const addSubSectionTemplate_CT2 = async (req, res) => {};
-const deleteSubSectionTemplate_CT2 = async (req, res) => {};
 
 export {
   createTemplate_CT2,
   getTemplate_CT2,
   getTemplates_CT2,
-  deleteSubSectionTemplate_CT2,
   deleteTemplate_CT2,
-  editTemplate_CT2,
-  getSubSectionTemplate_CT2,
-  addSubSectionTemplate_CT2,
+  editTemplate_CT2
 };
