+++ backend/index.js
@@ -27,15 +27,12 @@ ConectDB()
 app.use("/api/users",usersRoutes);
 app.use("/api/templates",templatesRoutes);
 app.use("/api/client",clientUserRoutes);//ct1 : perfil
-
-app.use("/api/templatesCT2",templatesCT2Routes);//ct2: signos-vitales
-app.use("/api/templatesCT3",templatesCT3Routes);//ct3: curaciones
-app.use("/api/templatesCT4",templatesCT4Routes);//ct4: vacunas
-app.use("/api/templatesCT7",templatesCT5Routes);//ct5: health care record
-app.use("/api/templatesCT6",templatesCT6Routes);//ct6: signos-vitales
-app.use("/api/templatesCT7",templatesCT7Routes);//ct7: Diuresis y deposiciones
-
-
+app.use("/cambio-de-posicion",templatesCT2Routes);
+app.use("curaciones",templatesCT3Routes);
+app.use("/vacunas",templatesCT4Routes);
+app.use("/visita-medica",templatesCT5Routes);
+app.use("/signos-vitales",templatesCT6Routes);
+app.use("/otros",templatesCT6Routes);
 
 const PORT = process.env.PORT || 4000;
 app.listen(PORT, ()=>{
