+++ backend/routes/templateCT5Routes.js
@@ -5,7 +5,7 @@ import {
   editTemplate_CT5,
   getTemplate_CT5,
   getTemplates_CT5,
-} from "../controllers/templateCT5Controller.js";
+} from "../controllers/templatesController/templateCT5Controller.js";
 import checkAuth from "../middleware/checkAuth.js";
 
 const router = express.Router();
@@ -14,9 +14,9 @@ router.route("/")
  .get(checkAuth, getTemplates_CT5)
  .post(checkAuth, createTemplate_CT5);
 
-router.route("/:id")
-  .get(checkAuth, getTemplate_CT5)
-  .put(checkAuth, editTemplate_CT5)
-  .delete(checkAuth, deleteTemplate_CT5);
+// router.route("/:id")
+//   .get(checkAuth, getTemplate_CT5)
+//   .put(checkAuth, editTemplate_CT5)
+//   .delete(checkAuth, deleteTemplate_CT5);
 
 export default router;
